Decompile scrcmd_pokemon_misc, 1
Also named some variables before GetMonData/SetMonData while I was looking for specifically the Golden Leaves since all instances looked unlabeled

+++ src/overlay_bug_contest.c
@@ -30,7 +30,7 @@ void BugContest_InitEncounters(BUGCONTEST *bugContest);
 void BugContest_RestoreParty_RetrieveCaughtPokemon(BUGCONTEST *bugContest);
 u16 BugContest_JudgePlayerMon(BUGCONTEST *bugContest, POKEMON *pokemon);
 
-BUGCONTEST *BugContest_new(FieldSystem *fsys, u8 weekday) {
+BUGCONTEST *BugContest_new(FieldSystem *fsys, u32 weekday) {
     BUGCONTEST *bugContest;
 
     bugContest = (BUGCONTEST *)AllocFromHeap(3, sizeof(BUGCONTEST));
